feat: For navigating to housing district locations check to see if any of the personal housing. Adjust TeleportCost.
Check the residential housing options for a better option when ward isn't important. If you need to get to a npc that's in all the housing zones then you just want the cheapest teleport option. Include residential teleport options when calculating teleport cost if the character is on the homeworld, otherwise calc based on the cost to teleport to the major town.